refactored a little bit
