+++ package.json
@@ -1,0 +1,11 @@
+{
+  "name": "wordle",
+  "version": "1.0.0",
+  "main": "index.js",
+  "repository": "https://github.com/ulrokx/wordleplusplus.git",
+  "author": "Ricky <richard.kirk790@gmail.com>",
+  "license": "MIT",
+  "devDependencies": {
+    "parcel": "^2.3.2"
+  }
+}
